+++ src/test/fixtures/TestHistory.sol
@@ -1,0 +1,9 @@
+contract TestHistory {
+  uint x;
+  constructor(uint _x) {
+    x = _x;
+  }
+  function setX(uint _x) {
+    x = _x;
+  }
+}