+++ src/test/rest.test.ts
@@ -59,7 +59,7 @@ describe("rest_7", function () {
           config,
           isAsync: true
         });
-      }, RestStatus.BAD_REQUEST, /Need one of: address, balance, minbalance, maxbalance, nonce, minnonce, maxnonce, maxnumber, code, index, codeHash, chainid/);
+      }, RestStatus.BAD_REQUEST, /Need one of: address, balance, minbalance, maxbalance, nonce, minnonce, maxnonce, maxnumber, code, index, codeHash, contractName, codePtrAddress, codePtrChainId, chainid, external, limit, offset, ignoreChain/);
     });
 
   });